ops(blog): improve audit_content checks and fix test discovery
